Shader Work

+++ SDEngine/Texture2D.h
@@ -6,7 +6,7 @@ class Texture2D {
 
 public:
 	/*Wrap Behaviour: GL_REPEAT or GL_CLAMP. Filter Behaviour: GL_LINEAR or GL_NEAREST*/
-	Texture2D(const std::string& FileName, unsigned int ExpectedComponent, GLint WrapBehaviour = GL_REPEAT, GLint FilterBehaviour = GL_LINEAR);
+	Texture2D(const std::string& FileName, unsigned int ExpectedComponent, GLint WrapBehaviour = GL_REPEAT, GLfloat FilterBehaviour = GL_LINEAR);
 	virtual ~Texture2D();
 
 	void Bind(unsigned int Unit);
